Recent pull
Files conflicting: Readme.m and hold_request.rb where resolved according to the newest versions in master branch. You'll have the most recent version of master.

+++ db/migrate/20190925031856_create_libraries.rb
@@ -5,7 +5,7 @@ class CreateLibraries < ActiveRecord::Migration[5.2]
       t.string :university
       t.string :location
       t.integer :max_days
-      t.integer :overdue_fines
+      t.float :overdue_fines
       # t.string :caller_locations
